requirements.txt
autograd==1.7.0
certifi==2024.12.14
chardet==4.0.0
compressai==1.1.5
contourpy==1.3.1
cycler==0.12.1
filelock==3.13.1
fonttools==4.55.5
fsspec==2024.2.0
huggingface-hub==0.27.1
idna==2.10
imageio==2.36.1
imageio-ffmpeg==0.6.0
Jinja2==3.1.3
kiwisolver==1.4.8
kornia==0.7.4
kornia_rs==0.1.8
lazy_loader==0.4
MarkupSafe==2.1.5
matplotlib==3.10.0
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
packaging==24.2
pandas==2.2.3
pillow==10.2.0
ptflops==0.7.4
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytorch-msssim==1.0.0
pytz==2024.2
PyYAML==6.0.2
requests==2.25.1
safetensors==0.5.2
scikit-image==0.25.0
scipy==1.15.0
six==1.17.0
sympy==1.13.1
thop==0.1.1.post2209072238
tifffile==2024.12.12
timm==1.0.14
torch==2.5.1+cu118
torchattacks==3.5.1
torchaudio==2.5.1+cu118
torchvision==0.20.1+cu118
tqdm==4.67.1
triton==3.1.0
typing_extensions==4.9.0
tzdata==2024.2
urllib3==1.26.20
